Make Rupees an integer in the contract
rupeesSchema was z.number(), which emits "type": "number" and generated `num`
on the Dart side — so every money field in the mobile client was a `num`, and
building a Rupees out of one did not compile.

MOBILE.md 4.3 is explicit that this must be an int: money in a floating-point
type is how ₹1 goes missing. `.int()` changes nothing for TypeScript, which
infers `number` either way; it changes the JSON Schema, which changes the Dart.

The test meant to catch this had passed all along. `expect(rate, isA<int>())`
inspects the runtime value, and 12500 decoded from JSON is an int whatever the
declaration says — so it asserted nothing. It now assigns to a typed local,
which does not compile unless the field really is declared int.

+++ packages/types/src/schema/common.ts
@@ -23,8 +23,20 @@ export const timestampSchema = z.string();
 /** ISO date without time, e.g. "2026-08-31" */
 export const dateOnlySchema = z.string();
 
-/** Amounts are whole rupees (INR). No paise anywhere on the platform. */
-export const rupeesSchema = z.number();
+/**
+ * Amounts are whole rupees (INR). No paise anywhere on the platform.
+ *
+ * `.int()` is load-bearing, and not for TypeScript — it infers `number` either
+ * way. It is what makes the JSON Schema say `"type": "integer"` rather than
+ * `"number"`, which is what makes the generated Dart say `int` rather than
+ * `num`. MOBILE.md §4.3: *"`Rupees` is an `int`. Make it a Dart `int`, never a
+ * `double`. Money in a floating-point type is how ₹1 goes missing."*
+ *
+ * Without it the mobile models typed every amount `num`, and the test meant to
+ * catch that passed anyway — `expect(rate, isA<int>())` inspects the runtime
+ * value, and `12500` parsed from JSON is an `int` whatever the static type says.
+ */
+export const rupeesSchema = z.number().int();
 
 /**
  * Every persisted record carries these. `deletedAt` is a soft delete:
